Deleted directory aiohttpdemo_polls and moved files to aiohttp-login-session
