--- aiohttpdemo_polls/README.md
@@ -1,3 +1,0 @@
-# aiohttp-login-session
-
-In this repositoy the python code should be implemented.
