--- aiohttpdemo_polls/routes.py
@@ -1,6 +1,0 @@
-from views import index
-
-
-def setup_routes(app):
-    app.router.add_get('/', index)
-
