--- aiohttpdemo_polls/views.py
@@ -1,5 +1,0 @@
-from aiohttp import web
-
-
-async def index(request):
-    return web.Response(text='Hello world!')
