+++ views.py
@@ -1,0 +1,14 @@
+from aiohttp import web
+
+
+async def index(request):
+    return web.Response(text='Hello world!')
+
+async def input1(request):
+    text = sa
+    return web.Response(text=text)
+
+
+
+sa = input("Enter your name: ")
+
